requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
psycopg[binary]==3.2.1
sqlalchemy==2.0.32
pgvector==0.3.5
python-multipart==0.0.9
pymupdf==1.24.9
pypdf==4.2.0
openai==1.40.0
numpy==2.0.1
tenacity==9.0.0
httpx==0.27.2